reporter/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flask-report-maker"
version = "0.1.0"
description = "AI-powered report generation web app using Flask and Ollama"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
requires-python = ">=3.8"
dependencies = [
    "Flask>=2.0",
    "SQLAlchemy>=1.4",
    "flask_sqlalchemy>=2.5",
    "pandas>=1.3",
    "plotly>=5.0",
    "PyPDF2>=2.0",
    "python-pptx>=0.6",
    "Pillow>=8.0",
    "openpyxl>=3.0",
    "python-docx>=0.8",
    "requests>=2.25",
    "beautifulsoup4>=4.9",
    "weasyprint>=53.0",
    "ollama>=0.1", # Ollama client library
    "pytesseract>=0.3"
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["instance*", "templates*"]

